JBIDE-16866 - as7/wf should not be twiddling deployment scanners
JBIDE-16866 cleanup
